bundle20_loader: Add flag to parse named branches from changeset
It's on by default for now.

I have found at least one repository which does not target the same
hashes as the api heads call do for the same branch.

There is a FIXME in the commit which explains it. To reproduce,
uncomment the fixme and use the sample repository
a23905-race-source-archive.zip

According to the doc, it's somehow a possible behavior as there exists
multiple branch models (`hg help glossary`):
- bookmark
- named branch
- branch
- anonymous branch

Related T329