HgLoaderFromDisk supports corrupted revisions
Summary:
When a repository has corrupted revision, the revision and its descendants
are not loaded.

This commit only deals with missing `filelogs` and configures the exclusion
system.

Missing `filelogs` are recoverable errors that should be skipped or
saved as `SkippedContent` but being missing the `SkippedContent` cannot
be calculated. This point is left for future commits.

Reviewers: #reviewers

Subscribers: ardumont, Alphare, vlorentz

Differential Revision: https://forge.softwareheritage.org/D4688