Ignore the repository's config
`HGRCPATH` only tells Mercurial to ignore the user's config files, but
some repositories have a `.hg/hgrc` file (only in the case that you copy
the files instead of cloning, if present) that is usually used for server-side
configuration. We want to ignore this, since it might affect loading
and ask for hooks that are not there or are otherwise annoying/dangerous,
for example.