loader: don't ignore closed branches
The existing util for getting a repo's branches skips closed branches
but did not leave any explanation why, either in the code or in the
commit message. I cannot think of a good reason for ignoring closed
branches, so we're removing this exception, which in turn fixes the
incremental issue detailed in T3336.

This has affected existing tests of the two repositories that had closed
branches. A test for the incremental behavior was added as well.