new-loader: fix incremental load of multiple heads
Some branches have multiple heads for a single branch (they can even be
closed heads). The SWH data model does not yet handle this, so we fix
the issue by asking a more precise question to the repository by
including all locals heads that are already stored as revisions.

We also test that this resolves the issue where the new loader would
always see the additional head as missing.