Capture missing revision <-> hgnode-id scenario in a xfail test
anomad-d is the `user`-`repository` name.

It's a repository which presents an anomaly in some undefined way yet. That anomaly
makes the ingestion fail.

It's something that's happening currently once in a while with the bitbucket ingestion.
This commit is just a preparatory work to analyze the problem either if i have time to
or if someone else wants to.

To analyze comment the xfail mark and let the test fail, then debug. That should fairly
help.

Related to T3584