Fix branch bookmark id format so ingestion can finish
For bookmarks, the ids format listed are not aligned with the rest of the code. It's
human readable id as bytes string instead of bytes string. As it's not what's expected
by the caller, this failed the build.

This commit adds the extra mapping to fix the issue

Related to T3584